third_party/blink/web_tests/css3/flexbox/flex-flow-auto-margins-no-available-space-assert.html

<!DOCTYPE html>
<style>
* {
    display: flex;
    padding-bottom: 20pt;
    min-height: 0.7%;
    margin-top: 6000%;
    flex-shrink: 0;
    flex-basis: 7000%;
}
</style>
<script src="../../resources/check-layout.js"></script>
<abbr data-expected-height=30210270>
    <input></input>
</abbr>
<p> crbug.com/380201: Don't shrink below border/padding when stretching children within a flexbox with no available space.</p>
<script>
    window.checkLayout('abbr');
</script>
